+++ Civilizations/src/main/java/com/kylantraynor/civilizations/shapes/Shape.java
@@ -33,8 +33,8 @@ public abstract class Shape implements Visualizable{
 	public abstract int getWidth();
 	public abstract int getHeight();
 	public abstract int getLength();
-	public abstract int getVolume();
-	public abstract int getArea();
+	public abstract long getVolume();
+	public abstract long getArea();
 	public abstract boolean isInside(double x, double y, double z);
 	public abstract String toString();
 	public abstract Location[] getBlockLocations();
